+++ leap_ec/simple.py
@@ -18,7 +18,8 @@ from leap_ec.real_rep.initializers import create_real_vector
 # Function ea_solve()
 ##############################
 def ea_solve(function, bounds, generations=100, pop_size=2,
-             mutation_std=1.0, maximize=False, viz=False, viz_ylim=(0, 1)):
+             mutation_std=1.0, maximize=False, viz=False, viz_ylim=(0, 1),
+             hard_bounds=True):
     """Provides a simple, top-level interfact that optimizes a real-valued
     function using a simple generational EA.
 
@@ -33,6 +34,8 @@ def ea_solve(function, bounds, generations=100, pop_size=2,
     :param float mutation_std: the width of the mutation distribution
     :param bool maximize: whether to maximize the function (else minimize)
     :param bool viz: whether to display a live best-of-generation plot
+    :param bool hard_bounds: if True, bounds are enforced at all times during
+        evolution; otherwise they are only used to initialize the population.
 
     :param (float, float) viz_ylim: initial bounds to use of the plots
         vertical axis
@@ -66,11 +69,17 @@ def ea_solve(function, bounds, generations=100, pop_size=2,
 
     """
 
+    if hard_bounds:
+        mutation_op = mutate_gaussian(std=mutation_std, hard_bounds=bounds,
+                        expected_num_mutations='isotropic')
+    else:
+        mutation_op = mutate_gaussian(std=mutation_std,
+                        expected_num_mutations='isotropic')
+
     pipeline = [
         ops.tournament_selection,
         ops.clone,
-        mutate_gaussian(std=mutation_std, hard_bounds=bounds,
-                        expected_num_mutations='isotropic'),
+        mutation_op,
         ops.uniform_crossover(p_swap=0.4),
         ops.evaluate,
         ops.pool(size=pop_size)